CHROMIUM: devfreq: rk3399: fix accounting when rate change fails
If we don't have enough time in vblank to set a new DDR frequency, we
timeout and don't change the frequency. Since we do not want to return
an error in this case to prevent log spam, we need to set the freq
pointer to prevent the devfreq framework from incorrectly thinking we
successfully went to the other frequency.

BUG=b:74260504
TEST=Check that the rate for the dpll in
/sys/kernel/debug/clk/clk_summary always matches
/sys/class/devfreq/dmc/cur_freq

Change-Id: Iee0c9043f6af01dfff02fed0ac68f573ffdf2299
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/952862
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
(cherry picked from commit c88963f0aa31dd2520e9401bdec1070d53ea38dc)
Reviewed-on: https://chromium-review.googlesource.com/974343