CHROMIUM: esdfs: Add support for special download folder
Adds mount options to allow the download folder to
map to a different location, with different a different
uid/gid. These id's are specified in the mounter's ns,
and do not need to be mapped into the base ns of the
esdfs mount.

BUG=b:63876697
TEST=Mount esdfs with dl_loc, dl_uid, dl_gid set

Change-Id: I6a5ac4e522e137d49f7507bbf488299074fd2007
Signed-off-by: Daniel Rosenberg <drosen@google.com>
Reviewed-on: https://chromium-review.googlesource.com/952867
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@google.com>