FROMLIST: Input: gpio-keys - add support for wakeup event action
Add support for specifying event actions to trigger wakeup when using
the gpio-keys input device as a wakeup source.

This would allow the device to configure when to wakeup the system. For
example a gpio-keys input device for pen insert, may only want to wakeup
the system when ejecting the pen.

Suggested-by: Brian Norris <briannorris@chromium.org>
Signed-off-by: Jeffy Chen <jeffy.chen@rock-chips.com>
(am from https://patchwork.kernel.org/patch/10273801/)
[additions by dtor]

BUG=b:72717057
TEST=pen ejection/insertion in S3 on kevin and scarlet

Change-Id: I74a22f8ba590149da22a84484e4fe390e2d1be74
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/947568
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>