FROMLIST: arm64: dts: rockchip: assign clock rate for ACLK_VIO
The ACLK_VIO is a parent clock used by a several children,
its suggested clock rate is 400MHz. Right now it gets 400MHz
because it sources from CPLL(800M) and divides by 2 after reset.
It's good not to rely on default values like this, so let's
explicitly set it.
NOTE: it's expected that at least one board may override cru node and
set the CPLL to 1.6 GHz. On that board it will be very important to be
explicit about aclk-vio being 400 MHz.

BUG=b:74220297
TEST=none
(am from https://patchwork.kernel.org/patch/10274987/)

Change-Id: Iad82094a01978553f047c26dc83533d5b7e7ee76
Signed-off-by: Shunqian Zheng <zhengsq@rock-chips.com>
Reviewed-on: https://chromium-review.googlesource.com/956906
Reviewed-by: Douglas Anderson <dianders@chromium.org>