CHROMIUM: watchdog: Add new arm_smc watchdog driver
This patch adds a stub watchdog driver that can be used on ARM systems
with a Secure Monitor firmware to forward watchdog operations to
firmware via a Secure Monitor Call. This may be useful for platforms
using TrustZone that want the Secure Monitor firmware to have the final
control over the watchdog.

BUG=b:74259526
TEST=Booted on Elm with appropriate firmware, played around manually
with watchdog enable/disable, letting it run out and setting timeout.

Change-Id: Id339eaa3ebd46e82a7efd7886a156b9483905bfb
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/976324
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>