BACKPORT: x86/i915: Avoid keeping waitboost active for signalling thread.
Once a client has requested a waitboost, we keep that waitboost active
until all clients are no longer waiting. This is because we don't
distinguish which waiter deserves the boost. However, with the advent of
fence signaling, the signaler threads appear as waiters to the RPS
interrupt handler. So instead of using a single boolean to track when to
keep the waitboost active, use a counter of all outstanding waitboosted
requests.

At this point, I have removed all vestiges of the rate limiting on
clients. Whilst this means that compositors should remain more fluid,
it also means that boosts are more prevalent. See commit b29c19b64528
("drm/i915: Boost RPS frequency for CPU stalls") for a longer discussion
on the pros and cons of both approaches.

A drawback of this implementation is that it requires constant request
submission to keep the waitboost trimmed (as it is now cancelled when the
request is completed). This will be fine for a busy system, but near
idle the boosts may be kept for longer than desired (effectively tens of
vblanks worstcase) and there is a reliance on rc6 instead.

v2: Remove defunct rps.client_lock

Reported-by: Micha Winiarski <michal.winiarski@intel.com>
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Micha Winiarski <michal.winiarski@intel.com>
Reviewed-by: Micha Winiarski <michal.winiarski@intel.com>
(cherry picked from commit 7b92c1bd0540b64f54d98331d67e57266f9343c4)

Conflicts:
    drivers/gpu/drm/i915/i915_gem.c
    drivers/gpu/drm/i915/i915_gem_request.c
    drivers/gpu/drm/i915/i915_irq.c
    drivers/gpu/drm/i915/intel_drv.h
    drivers/gpu/drm/i915/intel_pm.c
    (Change made to the code which does not exist.)

BUG=b:75363705
TEST=The GT frequency can be verified with the interface
        /sys/class/drm/card0/gt_cur_freq_mhz
     It helps in improving the power consumption during video playback.
     Verified on BigBuckBunny 1080p@30 fps video stream,
     With this patch the improvement of about ~200mW observed.

Change-Id: Ic224679d79c643f682bf515c4224f7c360e8c6cf
Signed-off-by: Megha Aggarwal <megha.aggarwal@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/967569
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>