CHROMIUM: config: turn on CONFIG_I2C_CHARDEV for Qualcomm
This was left off, but it's on everywhere else and Chrome OS expects
it in general.

NOTE: some boards have this as "=m" and others "=y".  In general "=m"
is preferred unless there's some urgent need for it to be "=y", like
if we need to do an I2C-based firmware device at bootup.  See
CL:422907 and CL:366000 as examples.

This is the equivalent of:
  echo 'CONFIG_I2C_CHARDEV=m' >> \
    chromeos/config/arm64/chromiumos-qualcomm.flavour.config
  ./chromeos/scripts/kernelconfig olddefconfig

...atop an already normalized config.

BUG=chromium:795946
TEST=Boot and "modprobe i2c-dev"; can use "i2cdetect"

Change-Id: I07911683f2ae7b0b17457fed53ee0451c98aa7e2
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/969594
Reviewed-by: Guenter Roeck <groeck@chromium.org>
(cherry picked from commit 2bed422da47dbedecf88ed0885d865fa59857702)
Signed-off-by: Matt Atwood <matthew.s.atwood@intel.corp-partner.google.com>