CHROMIUM: x86: Use LAHF/SAHF instead of PUSHF/POPF with clang to save/restore FLAGS
By default clang emits PUSHF/POPF instructions to save and restore
the FLAGS register. With these instructions all FLAGS bits are saved and
restored, including the Interrupt enable flag (IF) and the trap flag
(TF), which is problematic in the kernel. Modern x86 CPUs support the
LAHF/SAHF instructions, which only save/restore a subset of register bits
(OF, SF, ZF, AF, PF, CF).

Tell clang to use LAHF/SAHF instead of PUSHF/POPF. This is a temporary
workaround, the implementation of an actual fix (in clang) is in progress.

BUG=b:74467640
TEST=Build and boot on poppy or soraka
  dmesg | grep "sleeping function called from invalid context"
    => no output

Change-Id: Ie5bf76b291c67bc9741741f97779316e2687318a
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/985638
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>