UPSTREAM: drm/i915/psr: Nuke aux frame sync
eDP spec states that aux frame is required to do PSR2 selective
update but i915 don't fully implement it. It sends the aux frame
sync messages but the value is always zero as the GTC is not enabled
in driver.

Through tests was findout that pannels can do selective update when
the y-coordinate is also included in SDP, that is why it is required
to run PSR2 in i915.

A dummy value is not useful at all to sink, so removing everything
related to aux frame sync, if GTC is enabled we can bring this back.

Cc: Vathsala Nagaraju <vathsala.nagaraju@intel.com>
Acked-by: Rodrigo Vivi <rodrigo.vivi@intel.com>
Reviewed-by: Dhinakaran Pandiyan <dhinakaran.pandiyan@intel.com>
Signed-off-by: Jos Roberto de Souza <jose.souza@intel.com>
Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20180328223046.16125-3-jose.souza@intel.com
(cherry picked from commit 6ce9b78a7388c37b33293666aa6fc61c177046e6)
from v4.18-rc1
Signed-off-by: Tarun Vyas <tarun.vyas@intel.com>

BUG=b:70920211, b:109823283
TEST=No *Potential atomic update error* in dmesg
with external display connected and browser opened
and cursor blinking in address bar for PSR entry-exit
TEST=No other regressions on KBL/APL/BYT based devices

Change-Id: I3e5383ae9b0c80332271b31ba20a02291fba1374
Reviewed-on: https://chromium-review.googlesource.com/1135416
Commit-Ready: Nathan Ciobanu <nathan.d.ciobanu@intel.com>
Tested-by: Casey G Bowman <casey.g.bowman@intel.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>