CHROMIUM: net: add notifier_chain callback to inet_release
This patch adds a blocking notifier chain for inet_release that is
invoked when sockets are released.

The purpose of this callback chain is to offer a way for xt_qtaguid to
ensure a socket being released in inet_release can be properly untagged
without introducing an explicit dependency on an optional netfilter
module inside af_inet.

BUG=b:77126438
BUG=b:73424026
TEST=- scripts/checkpatch.pl
     - Android's kernel/tests/net/test/run_net_test.sh all_tests.sh
Change-Id: I057c2e0ccf981387e5866cebeb5a1087be8153b4
Signed-off-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-on: https://chromium-review.googlesource.com/989380