CHROMIUM: arm64: dts: rockchip: scarlet: Don't use pinctrl controller hogs for BT and WiFi pins
On Scarlet we use a hack involving pinctrl and GPIO driven (fake-)
regulators to reset the Bluetooth and WiFi controllers at boot. The
hack works like this: The BT_EN and WLAN_EN pins are configured as
pinctrl controller hogs to drive them low early in the boot process.
Further we have GPIO enabled fake-regulators that drive the
corresponding pin to high when the controller is enabled.

This scheme worked fine, until commit 3f0ad8ee1d4a ("pinctrl: Really
force states during suspend/resume") changed the behavior of pinctrl
controller hogs at resume. The change consists in reapplying the default
pin configuration at resume, which ends up driving BT/WLAN_EN pins to low.
Disabled Bluetooth and WiFi on its own would be annoying enough, but since
the system is not aware that the controllers are disabled it keeps
accessing them, which leads to system crashes on resume as reported
in b/77863555.

Instead of using controller hogs the pins could be tied to the fake-
regulator devices involved in the hack, which would avoid the problem.
However it turns out that pinctrl is not needed (directly) to drive the
pins low. When the regulators are registered they are disabled, in
consequence regulator_register() sets the pin to low when it requests
the GPIO.

With the pinctrl controller hogs the BT/WLAN_EN was deasserted for
+150ms, when relying on regulator_register() (or pinctrl settings tied
to the regulator) this time is reduced to about 60ms, which is still
more than enough to reset the controllers:

[    0.128412] DBG: setting GPIO 0.8 to 0
[    0.128639] DBG: set mux of pin 0.8 to 0
[    0.129072] DBG: set mux of pin 0.5 to 0
[    0.129252] rockchip-pinctrl pinctrl: DBG: setting pull of GPIO0-5 to 5
[    0.129479] DBG: setting GPIO 0.4 to 0
[    0.129673] DBG: set mux of pin 0.4 to 0
[    0.192842] DBG: setting GPIO 0.8 to 1
[    0.193086] DBG: setting GPIO 0.4 to 1

BUG=b:77863555
TEST=suspend dru and resume

Change-Id: I45e8abd7a4e4d060aa13d10fc2a9fd009755a1fe
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1009023
Reviewed-by: Brian Norris <briannorris@chromium.org>