BACKPORT: FROMLIST: media: platform: Add ChromeOS EC CEC driver
The ChromeOS Embedded Controller can expose a CEC bus, this patch add the
driver for such feature of the Embedded Controller.

This driver is part of the cros-ec MFD and will be add as a sub-device when
the feature bit is exposed by the EC.

The controller will only handle a single logical address and handles
all the messages retries and will only expose Success or Error.

The controller will be tied to the HDMI CEC notifier by using the platform
DMI Data and the i915 device name and connector name.

Signed-off-by: Neil Armstrong <narmstrong@baylibre.com>
Reviewed-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>
Reviewed-by: Hans Verkuil <hans.verkuil@cisco.com>
(am from https://patchwork.freedesktop.org/patch/226536/)

Conflicts:
	drivers/media/platform/Makefile
	- adapt to chromeos-4.4 version

BUG=chromium:827390
TEST=cros-ec-cec probes and can communicate with the EC for CEC packets
CQ-DEPEND=CL:1021815, CL:1029863, CL:1082215

Change-Id: Ib53196cf817cd78247123793af3ffac8e5e82c2d
Reviewed-on: https://chromium-review.googlesource.com/1021816
Commit-Ready: Neil Armstrong <narmstrong@baylibre.com>
Tested-by: Neil Armstrong <narmstrong@baylibre.com>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Benson Leung <bleung@chromium.org>
Reviewed-by: Sean Paul <seanpaul@google.com>