FROMLIST: media: cec-notifier: Get notifier by device and connector name
In non device-tree world, we can need to get the notifier by the driver
name directly and eventually defer probe if not yet created.

This patch adds a variant of the get function by using the device name
instead and will not create a notifier if not yet created.

But the i915 driver exposes at least 2 HDMI connectors, this patch also
adds the possibility to add a connector name tied to the notifier device
to form a tuple and associate different CEC controllers for each HDMI
connectors.

Signed-off-by: Neil Armstrong <narmstrong@baylibre.com>
Reviewed-by: Hans Verkuil <hans.verkuil@cisco.com>
(am from https://patchwork.freedesktop.org/patch/226526/)

BUG=chromium:827390
TEST=cros-ec-cec gets the i915 cec notifier and physical address

Change-Id: I72814a2955fd774c3b99f305ef1b4effa9c1b981
Reviewed-on: https://chromium-review.googlesource.com/1029863
Commit-Ready: Neil Armstrong <narmstrong@baylibre.com>
Tested-by: Neil Armstrong <narmstrong@baylibre.com>
Reviewed-by: Sean Paul <seanpaul@chromium.org>