HACK: Support reboot and poweroff through the PON module
KPDPWR + RESIN_N == shutdown
RESIN_N == warm reset

KPDPWR alone still does power up, so careful getting out of the
shutdown state, to not mistakenly power it back up
(deassert KPDPWR first).

It's believed that the long term solution is that this should be
done in the BIOS, so this patch only needs to be carried until
we have BIOS that does this.

BUG=b:78885565
TEST=Sequences work better.

Change-Id: I2876c050f4ef6ed206b41b92c24bb52f40e726b6
Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Signed-off-by: Alexandru M Stan <amstan@chromium.org>