CHROMIUM: iwl7000: mvm: disable EBS only in low latency mode
With the patch mentioned below, we disabled EBS completely if fragmented
EBS is not supported and we are in high traffic.  This is not correct,
because we still want to run non-fragmented EBS in high traffic to
prevent scanning on empty channels, even when fragmented EBS is not
supported.

Fix this by making the decision based on low latency mode only.

Signed-off-by: Ayala Beker <ayala.beker@intel.com>
iwl7000-tree: 351b8880e19ba5b51bc9fcf7a54a6bf497793db2