CHROMIUM: ext4: Hack to allow mounting partition with jbd2_csum_v3.
DO NOT MERGE TO NON-FACTORY BRANCHES.

In Chromium OS Factory flow, imaging flow and manufacturing line may need to
mount the stateful partition, which was prepared from test image.

Starting from M67, the stateful partition has been formatted with JBD2 checksum
v3 for journaling so we need to backport the change so older kernels may
access them.

This is a special hack only for factory branches otherwise we'll need to
either rebase kernel (and this is known to break on some older platforms
due to WiFi and BT stack changes) or cutting a new factory branch (which
will take factories weeks to months to re-stabilize).

BUG=b:72655843
TEST=None

Change-Id: I1499c5fbee5ee74657799302b40e36838cd9e7fb
Signed-off-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1031692