CHROMIUM: media: uvcvideo: Short-term fix for huddly camera failure bug
uvcvideo sends a halt signal to request a camera to stop sending frames
as bulk inputs. This halt signal puts Huddly camera into a bad state.
This CL adds a quirk that ensures uvcvideo does not send a HALT signal
to HuddlyGo cameras. This way, HuddlyGo camera can stop on its own, when
it realizes that the driver stopped reading frames. This fix was put
into CrOS kernel 3.14 in cl/717816 and we're doing the same for Teemo
CfM units with kernel 4.4 . Buganizer:b/65557515

BUG=chromium:839517
TEST=build kernel package and stress test HuddlyGo camera

Change-Id: I3bd8777d5a31397581c73b0cf769d4b54bfd6196
Reviewed-on: https://chromium-review.googlesource.com/1041049
Reviewed-by: Zhongze Hu <frankhu@google.com>
Commit-Queue: Dennis Wu <denniswu@google.com>
Tested-by: Dennis Wu <denniswu@google.com>