CHROMIUM: iwl7000: mvm: Send LQ command as async when necessary
The parameter that indicated whether the LQ command should be sent
as sync or async was removed, causing the LQ command to be sent as
sync from interrupt context (e.g. from the RX path). This resulted
in a kernel warning: "scheduling while atomic" and failing to send
the LQ command, which ultimately leads to a queue hang.

Fix it by adding back the required parameter to send the command as
sync only when it is allowed.

BUG=chromium:818846
BUG=b:78436635
TEST=test if connection with 4-way handshakes work

Change-Id: I54cd6c388cc9181ae26349c5ed36f4389d40f3a8
Signed-off-by: Avraham Stern <avraham.stern@intel.com>
iwl7000-tree: 13ff5b34ca605492b2d086a2727242fc0dd9a7b8
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/1074653
Commit-Ready: Kirtika Ruchandani <kirtika@chromium.org>
Tested-by: Kirtika Ruchandani <kirtika@chromium.org>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>