CHROMIUM: config: Select CONFIG_PCIEASPM_DEBUG=y
This makes available certain sysfs files that the user space can
use to debug and control the ASPM behavior. This is to use this
as a tweaking / debug tool on cyan.

(Slightly related: I recently sent a patch upstream to remove
this configuration altogether, and is waiting upstream review.
For more details:
https://patchwork.kernel.org/patch/10392591/)

Signed-off-by: Rajat Jain <rajatja@google.com>
BUG=b:79427560
TEST=Verify sysfs files

Change-Id: Iea0f0d155eb55f28d848735e36c1add2a322b75d
Reviewed-on: https://chromium-review.googlesource.com/1050974
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Rajat Jain <rajatja@chromium.org>
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>