CHROMIUM: media: rockchip-vpu: Stop zeroing registers for rk3399 VP8 decode
Not only the soft reset seems to be enough, but also the .dec_reg_num in
rk3399_vpu_variant struct did not cover all the VDEC registers. Stop
wasting CPU and bus cycles.

Since this was the only user of .dec_reg_num and .enc_reg_num was unused
already, remove both.

BUG=chromium:829739
TEST=Run WebRTC Multiple peer connections demo and Hangouts video call
 simultaneously on kevin

Change-Id: Ie04d73556eeab87344bacb390429a19532b081bb
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1051506
Reviewed-by: Alpha Lin <alpha.lin@rockchip.corp-partner.google.com>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>