FROMLIST: ASoC: Intel: kbl: Move codec sysclk config to codec_init function
patchwork link: https://patchwork.kernel.org/patch/10441443

On APL, commit fd0f237572ad
("ASoC: Intel: bxt: Move codec sysclk config to codec_init function")
fixed an issue related to jack detection.

The MCLK for DA7219 does not change in this platform, but is
currently being configured everytime as part of the platform_clock
event handler for DAPM. The upshot of this is that we have
unnecessary calls to this function, and it also means that if
a stream hasn't yet been started, DA7219 driver does not have the
correct MCLK rates programmed and so the HP detection feature does
not operate as expected.

The same fix is needed on KBL.

This patch rectifies this issue by moving the sysclk call to
codec_init function so it's only called once at initialisation.

Signed-off-by: Mac Chiang <mac.chiang@intel.com>
Acked-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
Acked-by: Adam Thomson <Adam.Thomson.Opensource@diasemi.com>

BUG=b:78436451
TEST=check HP/HEADSET/LINEOUT/KEYS detection work.

Change-Id: Idc5050a24bd038559dd8bd5a06574337921b3c28
Reviewed-on: https://chromium-review.googlesource.com/1053750
Commit-Ready: Mac Chiang <mac.chiang@intel.com>
Tested-by: Mac Chiang <mac.chiang@intel.com>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>