CHROMIUM: vsock: remove kiocb arg from dgram_dequeue in vsock_transport
This is a very targeted backport of upstream 1b784140474
"net: Remove iocb argument from sendmsg and recvmsg" only for the
vsock code.  The argument is not used and later vsock transport code
doesn't provide it.

I chose to do it this way because it touches much less
networking code and we're missing some of the preceding
infrastructure changes in 3.18 to do a clean cherry-pick.

This will break building with VMWARE_VMCI_VSOCKETS but we don't use
this and probably never will.

BUG=chromium:763970
TEST=build with USE=kvm_host on caroline

Signed-off-by: Sonny Rao <sonnyrao@chromium.org>

Change-Id: Ibb908e18a34e85e963987be6cd556fe69dda5c89
Reviewed-on: https://chromium-review.googlesource.com/1058939
Commit-Ready: Sonny Rao <sonnyrao@chromium.org>
Tested-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-by: Guenter Roeck <groeck@chromium.org>