UPSTREAM: ALSA: usb-audio: UAC3. Add support for mixer unit.
This adds support for the MIXER UNIT in UAC3. All the information
is obtained from the (HIGH CAPABILITY) Cluster's header. We don't
read the rest of the logical cluster to obtain the channel config
as that wont make any difference in the current mixer behaviour.

The name of the mixer unit is not yet requested as there is not
support for the UAC3 Class Specific String requests.

Tested in an UAC3 device working as a HEADSET with a basic mixer
unit (same as the one in the BADD spec) with no controls.

Signed-off-by: Jorge Sanjuan <jorge.sanjuan@codethink.co.uk>
Reviewed-by: Ruslan Bilovol <ruslan.bilovol@gmail.com>
Tested-by: Ruslan Bilovol <ruslan.bilovol@gmail.com>
Signed-off-by: Takashi Iwai <tiwai@suse.de>
(cherry picked from commit 6cfd839ae78ec3fac5ddbf7148155898727e90c3)

BUG=chromium:921092
TEST=With patch series, test audio jack on UAC3 USB dock

Change-Id: I6ca91b0d5bdca6845ccbac441f9fc1ae2dc9bb82
Signed-off-by: Cheng-Yi Chiang <cychiang@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1488695
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Tzung-Bi Shih <tzungbi@chromium.org>
(cherry picked from commit e6dcebc348ac580f1c67b7a367bd435383f5b89f)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1505072