UPSTREAM: drm/i915/psr: Fall back to max. synchronization latency if DPCD read fails
Noticed that we assume the best case of 0 latency when the DPCD read
fails, reasonable pessimism is safer.

eDP spec does say that if latency is greater than 8, the panel
supplier needs to provide it. I didn't see anything specific in the VBT
for this, so let's go with 8 frames as a fallback.

Cc: Jose Roberto de Souza <jose.souza@intel.com>
Cc: Ville Syrjl <ville.syrjala@linux.intel.com>
Reviewed-by: Jos Roberto de Souza <jose.souza@intel.com>
Signed-off-by: Dhinakaran Pandiyan <dhinakaran.pandiyan@intel.com>
Signed-off-by: Jani Nikula <jani.nikula@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20180511195145.3829-5-dhinakaran.pandiyan@intel.com
(cherry picked from commit 264ff016cf0672f99f98ea853dbc7e45c3527bc2)
from v4.17-rc3

BUG=chromium:870499
TEST=$(watch -c -d -n0.3 cat /run/debugfs_gpu/i915_edp_psr_status)
     and ensure that PSR is supported, enabled and active when there
     is no activity on the screen (eDP)

Signed-off-by: Nathan Ciobanu <nathan.d.ciobanu@intel.com>
Change-Id: I7fda5fc8a98018e2dec2328a87ec6609659a19d7
Reviewed-on: https://chromium-review.googlesource.com/1187477
Tested-by: Casey G Bowman <casey.g.bowman@intel.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>