BACKPORT: drm/i915/psr: Nuke PSR support for VLV and CHV
PSR hardware and hence the driver code for VLV and CHV deviates a lot from
their DDI counterparts. While the feature has been disabled for a long time
now, retaining support for these platforms is a maintenance burden. There
have been multiple refactoring commits to just keep the existing code for
these platforms in line with the rest. There are known issues that need to
be fixed to enable PSR on these platforms, and there is no PSR capable
platform in CI to ensure the code does not break again if we get around to
fixing the existing issues. On account of all these reasons, let's nuke
this code for now and bring it back if a need arises in the future.

Cc: Jani Nikula <jani.nikula@intel.com>
Cc: Rodrigo Vivi <rodrigo.vivi@intel.com>
Cc: Ville Syrjl <ville.syrjala@linux.intel.com>
Acked-by: Jani Nikula <jani.nikula@intel.com>
Acked-by: Rodrigo Vivi <rodrigo.vivi@intel.com>
Reviewed-by: Jos Roberto de Souza <jose.souza@intel.com>
Signed-off-by: Dhinakaran Pandiyan <dhinakaran.pandiyan@intel.com>
Signed-off-by: Jani Nikula <jani.nikula@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20180511230059.19387-1-dhinakaran.pandiyan@intel.com
(cherry picked from commit ce3508fd2a778e9366ab638f4e1dbe6dab874c5b)
from v4.17-rc3

Took this patch to remove a lot of PSR dead code (Nathan)

Conflicts:
	[Removed .has_psr=1 from VLV_FEATURES instead of
	resolving dependencies all the way back to eb6f771b4989 (Nathan)]
	drivers/gpu/drm/i915/i915_pci.c

	[Context delta (Nathan)]
	drivers/gpu/drm/i915/intel_psr.c

BUG=chromium:870499
TEST=$(watch -c -d -n0.3 cat /run/debugfs_gpu/i915_edp_psr_status)
     and ensure that PSR is supported, enabled and active when there
     is no activity on the screen (eDP)

Signed-off-by: Nathan Ciobanu <nathan.d.ciobanu@intel.com>
Change-Id: I3b6026b347443d67cc229579160c1eb8fa504d3b
Reviewed-on: https://chromium-review.googlesource.com/1187472
Tested-by: Casey G Bowman <casey.g.bowman@intel.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>