CHROMIUM: media: rockchip/isp1: Queue buffers in lockstep if two MIs are enabled
When both MIs have streaming enabled, both buffer queues need to operate
in lockstep, i.e. buffers at the same position in queues of both MIs need
to be used to output results of processing the same ISP frame.

Change the driver to have 2-stage internal queues, one for buffers
received for userspace for given MI and another for buffers ready for
producing output into. Push the buffers from the former to the latter
only if a pair is collected. Also use internal sequence ID counter to
match buffers of both streams together and allow streams to be
selectively stopped and started on-the-fly, without explicit
synchronization.

BUG=b:78779539
TEST=Make sure camera works in on scarlet.
CQ-DEPEND=CL:1066251

Change-Id: Ibef6faa1733440604a4b2d182195e8fa491a942d
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1059227