CHROMIUM: mkbp: Increase maximum mkbp event size
Having a 16 byte mkbp event size makes it possible to send CEC
messages from the EC to the AP directly inside the mkbp event
instead of first doing a notification and then a read.

This version of the change includes a fix for a failed
performance auto-test on bob/kevin. The problem was
that the previous version did not properly fall back
to v0 of the mkbp event since the return value that
it looked on was that of the transfer, not that of
the protocol error.

Signed-off-by: Stefan Adolfsson <sadolfsson@chromium.org>

BRANCH=none
TEST=Verified that bob falls back to v0 version of the mkbp
event and that Fizz works with both v0 and v1 (using some
extra debug prints). Also made sure there is nothing strange
when running "top" on bob, and that bobs keyboard works.
BUG=chromium:842873, b:76467407
Change-Id: I39e96cd17213f6b702707b3448859fafa682f146
Reviewed-on: https://chromium-review.googlesource.com/1061879
Commit-Ready: Stefan Adolfsson <sadolfsson@chromium.org>
Tested-by: Stefan Adolfsson <sadolfsson@chromium.org>
Reviewed-by: Stefan Adolfsson <sadolfsson@chromium.org>
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
(cherry picked from commit ed5018b2daae9fabbf95d819b6fb14c968372dc8)
Reviewed-on: https://chromium-review.googlesource.com/1071670
Reviewed-by: Daisuke Nojiri <dnojiri@chromium.org>
Commit-Queue: Stefan Adolfsson <sadolfsson@chromium.org>