CHROMIUM: iwl7000: mvm: honor the max_amsdu_subframes limit
A peer can limit the number of subframes it can handle in a
single A-MSDU.
Honor this limit.
Note that the smallest limit is 8, and we are very unlikely
to reach that limit. So this isn't really a big deal.

Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
automatic-review: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: 46a984117320385ad73da85310e61bbba305f2f3