CHROMIUM: android: binder: Fix invalid copy_from_user() calls
This patch fixes a couple of locations where copy_from_user() was being
called while preemption was disabled. This was leading to BUG()
messages. We instead use the copy_from_user_preempt_disabled() variant
which is defined in this file, and handles this scenario by enabling
preemption, but disabling rescheduling while the copy is taking place.

BUG=b:64854597
TEST=eve builds and boots fine, no BUG() messages seen.

Change-Id: Id6a67e5a534042afdbe25c0a7fd6ea2d9a442d1f
Signed-off-by: Prashant Malani <pmalani@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1065195
Reviewed-by: Dylan Reid <dgreid@chromium.org>