UPSTREAM: drm/mm: Reject over-sized allocation requests early
As we keep an rbtree of available holes sorted by their size, we can
very easily determine if there is any hole large enough that might
satisfy the allocation request. This helps when dealing with a highly
fragmented address space and a request for a search by address.

To cache the largest size, we convert into the cached rbtree variant
which tracks the leftmost node for us. However, currently we sorted into
ascending size order so the leftmost node is the smallest, and so to
make it the largest hole we need to invert our sorting.

BUG=b:112542653
TEST=boot grunt, run login_LoginSuccess, graphics_SanAngeles, cheets_StartAndroid

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Joonas Lahtinen <joonas.lahtinen@linux.intel.com>
Reviewed-by: Joonas Lahtinen <joonas.lahtinen@linux.intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20180521082131.13744-1-chris@chris-wilson.co.uk
(cherry picked from commit 2f7e87692e0441abf27a9714991edd136e87363a)
Signed-off-by: Dominik Behr <dbehr@chromium.org>

Change-Id: I4b087672fef320d37dcaad4f658dad4ca3fa0ce5