FROMLIST: usb: dwc2: fix the incorrect bitmaps for the ports of multi_tt hub
The dwc2_get_ls_map() use ttport to reference into the
bitmap if we're on a multi_tt hub. But the bitmaps index
from 0 to (hub->maxchild - 1), while the ttport index from
1 to hub->maxchild. This will cause invalid memory access
when the number of ttport is hub->maxchild.

Without this patch, I can easily meet a Kernel panic issue
if connect a low-speed USB mouse with the max port of FE2.1
multi-tt hub (1a40:0201) on rk3288 platform.

Signed-off-by: William Wu <william.wu@rock-chips.com>
Reviewed-by: Douglas Anderson <dianders@chromium.org>

BUG=None
TEST=Multi-TT hubs still work fine

Change-Id: I37efd00b63b2061a31ecd6ac6110883567ef02a4
Signed-off-by: Douglas Anderson <dianders@chromium.org>
(am from https://patchwork.kernel.org/patch/10414533/)
Reviewed-on: https://chromium-review.googlesource.com/1076323
Tested-by: William wu <wulf@rock-chips.com>
Reviewed-by: William wu <wulf@rock-chips.com>