CHROMIUM: config: Enable I2C_HID for w9013 digitizer
This is the equivalent of:

  echo 'CONFIG_I2C_HID=y' >> \
    chromeos/config/arm64/chromiumos-qualcomm.flavour.config
  ./chromeos/scripts/kernelconfig olddefconfig

...atop an already normalized config.

NOTE: This _doesn't_ change the "chromiumos-arm64" config, which had
this as "=m".  Argument here is that we will probably have the same
problem on real products as we saw in CL:366000 so we'll want this
'=y' for the targeted config.  However for the "multi" config we would
presumably be more OK with a delay loading things to keep the kernel
from getting too huge.

BUG=None
TEST=evtest and see "w9013 2D1F:002B" and events

Change-Id: I39de05a437b3f2acb2714744e9fbded8d8d2d675
Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1067663