CHROMIUM: drm/img-rogue/1.9: trigger GPU dump if FW fails to boot up
Request GPU debug dump in the GPU power-on code, if the firmware failed
to boot up properly.

The device power state is temporarily forced to ON, so the debug dump
will include a GPU register dump. By this point, the GPU should already
be powered and the register dump should be possible, even if the driver
has not set the power state to ON yet (this is usually done later in the
power-on sequence).

BUG=b:35648545
TEST=boot to UI and run graphics_GLBench

Change-Id: I76fab06c488ab37305f1e5a6d9c3c3a3f872f878
Signed-off-by: Afar Lin <afar.lin@imgtec.com>
Reviewed-on: https://chromium-review.googlesource.com/1068708
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Bernie Thompson <bhthompson@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>