CHROMIUM: config: Enable Intel Virtual Button Driver
Enable the Intel Virtual Button Driver.  There will be a later change
on some devices to replace the Google TBMC device with the Intel
Virtual Buttons device.  The TBMC device is a subset of functionality,
and by switching to the device we can eventually get rid of the
need to support and maintain a separate driver for the TBMC device.

BUG=None
TEST=Build and boot.  Verified on eve that entering and exiting
tablet mode is properly detected (even when adding in temporary
changes to disable the EC's functionality to also report tablet
mode via simulated keyboard events).

Signed-off-by: Matt Delco <delco@chromium.org>
Change-Id: Iac2ebfaec077f65936d107ba0f2f4f2eb61b97f6
Reviewed-on: https://chromium-review.googlesource.com/1073745
Commit-Ready: Matt Delco <delco@google.com>
Tested-by: Matt Delco <delco@google.com>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>